move train val epoch functions to separate engine module
